Fix code review issues: unsafe casts, NPE risks, magic numbers, and typos
Agent-Logs-Url: https://github.com/microsphere-projects/microsphere-spring-boot/sessions/2237cd14-7911-4a40-8f1a-5e1a275fd6fc

+++ microsphere-spring-boot-compatible/src/main/java/org/springframework/boot/web/servlet/MultipartConfigFactory.java
@@ -77,7 +77,7 @@ public class MultipartConfigFactory {
 		long maxRequestSizeBytes = convertToBytes(this.maxRequestSize, -1);
 		long fileSizeThresholdBytes = convertToBytes(this.fileSizeThreshold, 0);
 		return new MultipartConfigElement(this.location, maxFileSizeBytes, maxRequestSizeBytes,
-				(int) fileSizeThresholdBytes);
+				Math.toIntExact(fileSizeThresholdBytes));
 	}
 
 	/**
